Modernize packaging (Closes #4)
