+++ .editorconfig
@@ -1,9 +1,5 @@
-# Generated from:
-# https://github.com/plone/meta/tree/master/config/default
-# See the inline comments on how to expand/tweak this configuration file
-#
 # EditorConfig Configuration file, for more details see:
-# http://EditorConfig.org
+# https://EditorConfig.org
 # EditorConfig is a convention description, that could be interpreted
 # by multiple editors to enforce common coding conventions for specific
 # file types
@@ -29,11 +25,7 @@ max_line_length = off
 # 4 space indentation
 indent_size = 4
 
-[*.{yml,zpt,pt,dtml,zcml}]
-# 2 space indentation
-indent_size = 2
-
-[*.{json,jsonl,js,jsx,ts,tsx,css,less,scss,html}]  # Frontend development
+[*.{html,dtml,pt,zpt,xml,zcml,js,json,less,css,yml,yaml}]
 # 2 space indentation
 indent_size = 2
 
@@ -42,12 +34,3 @@ indent_size = 2
 indent_style = tab
 indent_size = unset
 tab_width = unset
-
-
-##
-# Add extra configuration options in .meta.toml:
-#  [editorconfig]
-#  extra_lines = """
-#  _your own configuration lines_
-#  """
-##
