+++ .github/workflows/main.yml
@@ -1,0 +1,61 @@
+name: collective.contact_behaviors CI
+
+on:
+  push:
+
+jobs:
+
+  config:
+    name: "Compute configuration values"
+    uses: ./.github/workflows/config.yml
+  lint:
+    name: "Lint codebase"
+    uses: plone/meta/.github/workflows/backend-lint.yml@2.x
+    needs:
+      - config
+    with:
+      python-version: ${{ needs.config.outputs.python-version }}
+      plone-version:  ${{ needs.config.outputs.plone-version }}
+  test:
+    name: "Test codebase"
+    uses: plone/meta/.github/workflows/backend-pytest.yml@2.x
+    needs:
+      - config
+    strategy:
+      matrix:
+        python-version: ["3.13", "3.12", "3.11", "3.10"]
+        plone-version: ["6.1-latest", "6.0-latest"]
+    with:
+      python-version: ${{ matrix.python-version }}
+      plone-version: ${{ matrix.plone-version }}
+
+  coverage:
+    name: "Backend: Coverage"
+    uses: plone/meta/.github/workflows/backend-pytest-coverage.yml@2.x
+    needs:
+      - config
+      - test
+    with:
+      python-version: ${{ needs.config.outputs.python-version }}
+      plone-version:  ${{ needs.config.outputs.plone-version }}
+
+  report:
+    name: "Final report"
+    if: ${{ always() }}
+    runs-on: ubuntu-latest
+    needs:
+      - config
+      - lint
+      - test
+      - coverage
+    steps:
+      - name: Report
+        shell: bash
+        run: |
+          echo '# Workflow Report' >> $GITHUB_STEP_SUMMARY
+          echo '| Job ID | Conclusion |' >> $GITHUB_STEP_SUMMARY
+          echo '| --- | --- |' >> $GITHUB_STEP_SUMMARY
+          echo '| Config | ${{ needs.config.result }} |' >> $GITHUB_STEP_SUMMARY
+          echo '| Lint | ${{ needs.coverage.result }} |' >> $GITHUB_STEP_SUMMARY
+          echo '| Test | ${{ needs.coverage.result }} |' >> $GITHUB_STEP_SUMMARY
+          echo '| Coverage | ${{ needs.coverage.result }} |' >> $GITHUB_STEP_SUMMARY
