--- .github/workflows/meta.yml
@@ -1,28 +1,0 @@
-# Generated from:
-# https://github.com/plone/meta/tree/master/config/default
-# See the inline comments on how to expand/tweak this configuration file
-name: Meta
-on:
-  push:
-    branches:
-      - master
-      - main
-  pull_request:
-    branches:
-      - master
-      - main
-  workflow_dispatch:
-
-jobs:
-  qa:
-    uses: plone/meta/.github/workflows/qa.yml@master
-  test:
-    uses: plone/meta/.github/workflows/test.yml@master
-  coverage:
-    uses: plone/meta/.github/workflows/coverage.yml@master
-  dependencies:
-    uses: plone/meta/.github/workflows/dependencies.yml@master
-  release-ready:
-    uses: plone/meta/.github/workflows/release_ready.yml@master
-  circular:
-    uses: plone/meta/.github/workflows/circular.yml@master
