+++ .gitignore
@@ -1,53 +1,18 @@
-# Generated from:
-# https://github.com/plone/meta/tree/master/config/default
-# See the inline comments on how to expand/tweak this configuration file
-# python related
+!src/pas
+.coverage*
+.pydevproject
+.pytest_cache
+.ruff_cache
+.python-version
+.venv
 *.egg-info
+*.log
+*.mo
 *.pyc
-*.pyo
-
-# tools related
-build/
-.coverage
-coverage.xml
-dist/
-docs/_build
-__pycache__/
-.tox
-.vscode/
-node_modules/
-
-# venv / buildout related
-bin/
-develop-eggs/
-eggs/
-.eggs/
-etc/
-.installed.cfg
-include/
-lib/
-lib64
-.mr.developer.cfg
-parts/
-pyvenv.cfg
-var/
-
-# mxdev
-/instance/
-/.make-sentinels/
-/*-mxdev.txt
-/reports/
-/sources/
-/venv/
-.installed.txt
-
+*.swp
+constraints-mxdev.txt
+constraints.txt
+instance
 requirements-mxdev.txt
-!*.mo
-
-##
-# Add extra configuration options in .meta.toml:
-#  [gitignore]
-#  extra_lines = """
-#  _your own configuration lines_
-#  """
-##
+requirements.lock
+requirements.txt
