--- .meta.toml
@@ -1,31 +1,0 @@
-# Generated from:
-# https://github.com/plone/meta/tree/master/config/default
-# See the inline comments on how to expand/tweak this configuration file
-[meta]
-template = "default"
-commit-id = "23d5b8e1"
-
-[pyproject]
-codespell_skip = "*.min.js,*.pot,*.po,*.yaml,*.json"
-codespell_ignores = "vew"
-dependencies_ignores = "['plone.volto', 'zestreleaser.towncrier', 'zest.releaser', 'pytest', 'pytest-cov', 'pytest-plone']"
-dependencies_mappings = [
-    "Plone = ['Products.CMFPlone', 'Products.CMFCore', 'Products.GenericSetup']",
-    ]
-check_manifest_ignores = """
-    "news/*",
-    "requirements-mxdev.txt",
-"""
-extra_lines = """
-ignore-bad-ideas = ["*.mo"]
-"""
-
-[gitignore]
-extra_lines = """
-requirements-mxdev.txt
-!*.mo
-"""
-
-[tox]
-test_runner = "pytest"
-test_path = "/tests"
