--- .pre-commit-config.yaml
@@ -1,84 +1,0 @@
-# Generated from:
-# https://github.com/plone/meta/tree/master/config/default
-# See the inline comments on how to expand/tweak this configuration file
-ci:
-    autofix_prs: false
-    autoupdate_schedule: monthly
-
-repos:
--   repo: https://github.com/asottile/pyupgrade
-    rev: v3.4.0
-    hooks:
-    -   id: pyupgrade
-        args: [--py38-plus]
--   repo: https://github.com/pycqa/isort
-    rev: 5.12.0
-    hooks:
-    -   id: isort
--   repo: https://github.com/psf/black
-    rev: 23.3.0
-    hooks:
-    -   id: black
--   repo: https://github.com/collective/zpretty
-    rev: 3.1.0a2
-    hooks:
-    -   id: zpretty
-
-##
-# Add extra configuration options in .meta.toml:
-#  [pre_commit]
-#  zpretty_extra_lines = """
-#  _your own configuration lines_
-#  """
-##
--   repo: https://github.com/PyCQA/flake8
-    rev: 6.0.0
-    hooks:
-    -   id: flake8
-
-##
-# Add extra configuration options in .meta.toml:
-#  [pre_commit]
-#  flake8_extra_lines = """
-#  _your own configuration lines_
-#  """
-##
--   repo: https://github.com/codespell-project/codespell
-    rev: v2.2.4
-    hooks:
-    -   id: codespell
-        additional_dependencies:
-          - tomli
-
-##
-# Add extra configuration options in .meta.toml:
-#  [pre_commit]
-#  codespell_extra_lines = """
-#  _your own configuration lines_
-#  """
-##
--   repo: https://github.com/mgedmin/check-manifest
-    rev: "0.49"
-    hooks:
-    -   id: check-manifest
--   repo: https://github.com/regebro/pyroma
-    rev: "4.2"
-    hooks:
-    -   id: pyroma
--   repo: https://github.com/mgedmin/check-python-versions
-    rev: "0.21.2"
-    hooks:
-    -   id: check-python-versions
-        args: ['--only', 'setup.py,pyproject.toml']
--   repo: https://github.com/collective/i18ndude
-    rev: "6.0.0"
-    hooks:
-    -   id: i18ndude
-
-##
-# Add extra configuration options in .meta.toml:
-#  [pre_commit]
-#  extra_lines = """
-#  _your own configuration lines_
-#  """
-##
