--- constraints.txt
@@ -1,1 +1,0 @@
--c https://dist.plone.org/release/6.0-latest/constraints.txt
