+++ instance.yaml
@@ -1,8 +1,3 @@
 default_context:
-    initial_user_name: 'admin'
     initial_user_password: 'admin'
-
-    load_zcml:
-        package_includes: ['collective.contact_behaviors']
-
-    db_storage: direct
+    zcml_package_includes: 'collective.contact_behaviors'
