+++ mx.ini
@@ -4,6 +4,7 @@
 ; to learn more about mxdev visit https://pypi.org/project/mxdev/
 
 [settings]
+main-package = -e .[test]
 ; example how to override a package version
 ; version-overrides =
 ;     example.package==2.1.0a2
@@ -14,3 +15,8 @@
 ; pushurl = git@github.com:collective/example.contenttype.git
 ; extras = test
 ; branch = feature-7
+version-overrides =
+    plone.restapi>=9.10.0
+    pytest-plone>=1.0.0a1
+    plone.autoinclude>=2.0.3
+    zest.releaser>=9.5.0
