+++ news/4.internal
@@ -1,0 +1,1 @@
+Modernize packaging. @ericof
