+++ requirements.txt
@@ -1,1 +1,1 @@
--e ".[test]"
+-c https://dist.plone.org/release/6.1.2/constraints.txt
