+++ src/collective/contact_behaviors/__init__.py
@@ -1,8 +1,11 @@
 """Init and utils."""
+
 from zope.i18nmessageid import MessageFactory
 
 import logging
 
+
+__version__ = "1.0.0a4.dev0"
 
 PACKAGE_NAME = "collective.contact_behaviors"
