+++ tests/vocabularies/test_vocab_countries.py
@@ -30,7 +30,7 @@ class TestVocabAvailableCountries:
         ],
     )
     def test_token(self, token):
-        assert token in [x for x in self.vocab.by_token]
+        assert token in list(self.vocab.by_token)
 
     @pytest.mark.parametrize(
         "token,title",
@@ -72,7 +72,7 @@ class TestVocabCountries:
         ],
     )
     def test_token(self, token):
-        assert token in [x for x in self.vocab.by_token]
+        assert token in list(self.vocab.by_token)
 
     @pytest.mark.parametrize(
         "token",
@@ -82,7 +82,7 @@ class TestVocabCountries:
         ],
     )
     def test_token_not_in(self, token):
-        assert token not in [x for x in self.vocab.by_token]
+        assert token not in list(self.vocab.by_token)
 
     @pytest.mark.parametrize(
         "token,title",
